requirements.txt
jinja2==2.8
dnspython==1.14.0